--- src/s1/t2/n1/exercici1/Exceptions.java
@@ -1,8 +1,0 @@
-package s1.t2.n1.exercici1;
-
-
-class VendaBuidaException extends Exception {
-    VendaBuidaException (){
-        super("Per fer una venda primer has d'afegir productes");
-    }
-}
