+++ src/s1/t2/n1/exercici1/classes/Producte.java
@@ -1,0 +1,16 @@
+package s1.t2.n1.exercici1.classes;
+
+public class Producte {
+    private float preu;
+    public String nom;
+
+
+    public Producte(float preu, String nom) {
+        this.preu = preu;
+        this.nom = nom;
+    }
+
+    public float getPreu () {
+        return preu;
+    }
+}
