+++ src/s1/t2/n1/exercici1/exceptions/VendaBuidaException.java
@@ -1,0 +1,8 @@
+package s1.t2.n1.exercici1.exceptions;
+
+
+public class VendaBuidaException extends Exception {
+    public VendaBuidaException (){
+        super("Per fer una venda primer has d'afegir productes");
+    }
+}
